requirements.txt
# Scraping (Phase 1)
google-play-scraper==1.2.7

# LLM
groq==0.11.0

# Config
python-dotenv==1.0.1

# Data validation
pydantic==2.7.1

# PII detection (Phase 2)
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
spacy==3.7.4
# After install: python -m spacy download en_core_web_lg

# Email templating (Phase 4)
jinja2==3.1.4

# Resilience
tenacity==8.3.0

# CLI
click==8.1.7

# Web UI (Phase 5)
streamlit==1.35.0

# Testing
pytest==8.2.2
pytest-mock==3.14.0

# Scheduler
apscheduler==3.10.4
